+++ lib/rspock/ast/then_block.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 require 'rspock/ast/block'
 require 'rspock/ast/comparison_to_assertion_transformation'
+require 'rspock/ast/interaction_transformation'
 
 module RSpock
   module AST
@@ -14,7 +15,19 @@ module RSpock
       end
 
       def children
-        super.map { |child| ComparisonToAssertionTransformation.new(:_test_index_, :_line_number_).run(child) }
+        super.reject { |child| interaction_transformation.interaction_node?(child) }
+          .map { |child| ComparisonToAssertionTransformation.new(:_test_index_, :_line_number_).run(child) }
+      end
+
+      def interactions
+        @children.select { |child| interaction_transformation.interaction_node?(child) }
+          .map { |child| interaction_transformation.run(child) }
+      end
+
+      private
+
+      def interaction_transformation
+        @interaction_transformation ||= RSpock::AST::InteractionTransformation.new
       end
     end
   end