+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED]
+### Added
+- Interaction-based testing: Mock with expectations in the Then block.
+
 ### Changed
 - Test names now have the test index and line number as suffix instead of prefix.
 - Cleanup transformed code output.